requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.74
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
transformers==4.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0